🔧Refactor modal screens to use new base classes for improved exit and refresh handling
