+++ src/sourcerer/presentation/screens/provider_creds_registration/main.py
@@ -1,13 +1,10 @@
 from dataclasses import dataclass
 from enum import Enum
-from typing import ClassVar
 
 from dependency_injector.wiring import Provide
 from textual import on
 from textual.app import ComposeResult
-from textual.binding import Binding, BindingType
 from textual.containers import Container, Horizontal, VerticalScroll
-from textual.screen import ModalScreen
 from textual.widgets import Label, Select
 
 from sourcerer.domain.access_credentials.services import (
@@ -22,6 +19,7 @@ from sourcerer.infrastructure.access_credentials.registry import (
 )
 from sourcerer.infrastructure.utils import generate_unique_name
 from sourcerer.presentation.di_container import DiContainer
+from sourcerer.presentation.screens.shared.modal_screens import ExitBoundModalScreen
 from sourcerer.presentation.screens.shared.widgets.button import Button
 from sourcerer.presentation.screens.shared.widgets.labeled_input import LabeledInput
 
@@ -38,12 +36,8 @@ class ProviderCredentialsEntry:
     fields: dict[str, str]
 
 
-class ProviderCredsRegistrationScreen(ModalScreen):
+class ProviderCredsRegistrationScreen(ExitBoundModalScreen):
     CSS_PATH = "styles.tcss"
-
-    BINDINGS: ClassVar[list[BindingType]] = [
-        Binding("escape", "cancel_screen", "Pop screen"),
-    ]
 
     MAIN_CONTAINER_ID = "ProviderCredsRegistrationScreen"
     SETTINGS_CONTAINER_ID = "settings"
@@ -290,6 +284,3 @@ class ProviderCredsRegistrationScreen(ModalScreen):
         self.query_one(f"#{self.CREDENTIALS_FIELDS_CONTAINER_ID}").query_one(
             ".form_input"
         ).focus()
-
-    def action_cancel_screen(self):
-        self.dismiss()
